requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=9.0.0
tqdm>=4.64.0
wandb>=0.13.0
opencv-python>=4.5.0
diffusers>=0.30.0